requirements.txt
Package                 Version
----------------------- -----------
absl-py                 1.4.0
aiohttp                 3.8.4
aiosignal               1.3.1
albumentations          1.3.0
anyio                   3.6.2
appdirs                 1.4.4
arrow                   1.2.3
async-timeout           4.0.2
attrs                   22.2.0
beautifulsoup4          4.12.2
blessed                 1.20.0
brotlipy                0.7.0
cachetools              5.3.0
certifi                 2022.12.7
cffi                    1.15.1
charset-normalizer      2.0.4
click                   8.1.3
cloudpickle             2.2.1
colorama                0.4.6
contourpy               1.0.7
croniter                1.3.14
cryptography            39.0.1
cycler                  0.11.0
dataclasses             0.6
dateutils               0.6.12
deepdiff                6.3.0
docker-pycreds          0.4.0
docstring-parser        0.15
einops                  0.6.0
fastapi                 0.88.0
fastcore                1.5.29
fastdownload            0.0.7
fastprogress            1.0.3
filelock                3.9.0
flit_core               3.8.0
fonttools               4.39.2
frozenlist              1.3.3
fsspec                  2023.3.0
gitdb                   4.0.10
GitPython               3.1.31
gmpy2                   2.1.2
google-auth             2.16.2
google-auth-oauthlib    0.4.6
grpcio                  1.51.3
h11                     0.14.0
idna                    3.4
imageio                 2.26.0
importlib-metadata      6.1.0
importlib-resources     5.12.0
inquirer                3.1.3
itsdangerous            2.1.2
Jinja2                  3.1.2
joblib                  1.2.0
jsonargparse            4.21.0
kaggle                  1.5.13
kiwisolver              1.4.4
kornia                  0.6.12
lazy_loader             0.1
lightning               2.0.2
lightning-cloud         0.5.34
lightning-flash         0.8.1.post0
lightning-utilities     0.8.0
llvmlite                0.39.1
Markdown                3.4.1
markdown-it-py          2.2.0
MarkupSafe              2.1.1
matplotlib              3.7.1
mdurl                   0.1.2
mkl-fft                 1.3.1
mkl-random              1.2.2
mkl-service             2.4.0
mpmath                  1.2.1
multidict               6.0.4
munkres                 1.1.4
networkx                2.8.4
numba                   0.56.4
numpy                   1.23.5
oauthlib                3.2.2
opencv-python           4.7.0.72
opencv-python-headless  4.7.0.72
ordered-set             4.1.0
packaging               22.0
pandas                  1.5.3
pathtools               0.1.2
Pillow                  9.4.0
pip                     23.0.1
plotly                  5.13.1
protobuf                3.20.1
psutil                  5.9.4
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pycparser               2.21
pydantic                1.10.6
pyDeprecate             0.3.2
Pygments                2.15.0
PyJWT                   2.6.0
pynndescent             0.5.8
pyOpenSSL               23.0.0
pyparsing               3.0.9
PySocks                 1.7.1
pythae                  0.1.1
python-dateutil         2.8.2
python-editor           1.0.4
python-multipart        0.0.6
python-slugify          8.0.1
pytorch-lightning       1.9.0
pytz                    2022.7.1
PyWavelets              1.4.1
PyYAML                  6.0
qudida                  0.0.4
readchar                4.0.5
requests                2.28.1
requests-oauthlib       1.3.1
rich                    13.3.4
rsa                     4.9
scikit-image            0.20.0
scikit-learn            1.2.1
scipy                   1.9.1
sentry-sdk              1.17.0
setproctitle            1.3.2
setuptools              59.5.0
six                     1.16.0
smmap                   5.0.0
sniffio                 1.3.0
soupsieve               2.4
starlette               0.22.0
starsessions            1.3.0
sympy                   1.11.1
tenacity                8.2.2
tensorboard             2.12.0
tensorboard-data-server 0.7.0
tensorboard-plugin-wit  1.8.1
text-unidecode          1.3
threadpoolctl           3.1.0
tifffile                2023.3.15
torch                   2.0.0
torchaudio              2.0.0
torchmetrics            0.10.3
torchvision             0.15.0
tqdm                    4.65.0
traitlets               5.9.0
triton                  2.0.0
typeshed-client         2.2.0
typing_extensions       4.4.0
umap-learn              0.5.3
urllib3                 1.26.14
uvicorn                 0.21.1
wandb                   0.14.0
wcwidth                 0.2.6
websocket-client        1.5.1
websockets              11.0.1
Werkzeug                2.2.3
wheel                   0.38.4
yarl                    1.8.2
zipp                    3.15.0